refactoring ui
