+++ src/app/context/PageContext.tsx
@@ -1,0 +1,7 @@
+// src/app/(site)/dashboard/page.tsx
+"use client";
+
+import { createContext } from "react";
+import { Context } from "interfaces";
+
+export const PageContext = createContext<Context>({});
