+++ tsconfig.json
@@ -20,7 +20,8 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "components/*": ["./src/app/components/*"]
+      "components/*": ["./src/app/components/*"],
+      "interfaces" : ["./src/app/global"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
